Performance improvements when NIC Enabled
Gathering all the network interfaces can be expensive:

```
NetworkInterface.GetAllNetworkInterfaces()
```

That expression can take around 100ms on my machine. Thus calling it every
interval can lead to some noticeable delays. The fix is to hook into an
event that is only called when an interface changes, and then we'd refresh
the internal list.

The NetworkChange.NetworkAddressChanged event caters to our needs.

Switching to this event eliminated the performance degradation (except on
startup and when an interface changes -- but we have to pay the price
somewhere).

Also the logic for detecting a change in interfaces was flawed as one
could create and destroy a virtual NIC in the interval number of seconds
and the previous logic wouldn't detect a change (as the total number of
interfaces wouldn't change).