Add prompt for web server restart
Web server can be checked if it needs to be restarted by checking contents of PendingRestarts HashSet<PendingRestartReason> or RestartRequired bool.

Added events to inform when pending restart reason is added and when running state is updated. Will be useful in future to create indicator or notification/reminder.

Auth and port forms prompt you to restart web server if needed.

When run for first time or when conifguration for user/pass for http auth is missing it will be filled with "librehm" in auth form for first time.

If password hash is set(contains 64chars) then pass textbox in auth form will also be filled with PassChars.

Pressing enter in textboxes on authform jumps to next control.