Fix ASUS X470-I incontrollable fans
I had correct fan readings for the fan on the MB, but could not control them. Found in the code there is an offset of 2; the comment states that the first 2 fans always show zero. I removed the offset - now I have 5 fan controls (which is not correct, since the MB only has 3 fan headers), but the first 2 actually control the fans.
I assume that the offset is either outdated and no longer needed, or it should offset 2 fans from the end of the array, not the start of it. I don't have all the headers populated, so cannot fully test whether the offset from the end is fine.

I built the LHM library and added it to both LHM and FanCtrl v1.5.6, in both cases I could control the fans.