Reinstate GenericMemory with unambiguous identifier
Restores the GenericMemory class to fix the duplicate hardware
identifier ("ram"), introduced by the recent split into separate
TotalMemory and VirtualMemory classes

The alternative would be to have ram hardware instances (which imo
doesn't make sense as it's the same hardware) or have them both use
different names.

On the plus side this cuts the number of calls to `GlobalMemoryStatusEx`
in half ;)

And I'm biased against name changes as tooling can depend on names, so
the 6 sensors have had their names reverted back to their original
prefixes of "Memory" for physical memory and "Virtual Memory" for ...
virtual memory.