Support ASUS ROG Crosshair X870E Apex (#1834)
* Support ASUS ROG Crosshair X870E Apex

* Remove unnecessary HashSet

* Adjust Temperature Sensor Mapping

Both Temp #1 and Temp #22 move with CPU load, but it looks like Temp #22 is actually CPU which is the same as the Hero. #1 doesn't align mostly with VRM but isn't an exact match.

* Refactor temperature sensor mapping

Replaces the temperature sensor dictionary mapping with explicit additions for known sensors and a loop to add unmapped sensors.